mkfs.gfs2: Move the new rgrp creation code into libgfs2
Adapt and move the new resource group creation and writing code into
libgfs2. The new APIs have been created to no longer expose the
underlying container (i.e. osi_tree.h bits) to users, and to reduce
dependency on the gfs2_sbd structure which has traditionally been the
dumping ground for long-lived state.

It is intended to convert the rest of the mkfs code to use these APIs
and then migrate the other tools to use them gradually and eventually
hide the rgrp_tree structure and remove the old functions and structures
which the new ones obsolete.

Signed-off-by: Andrew Price <anprice@redhat.com>