fsck.gfs2: Don't rely on cluster.conf when rebuilding sb
As cluster.conf no longer exists we can't sniff the locking options from
it when rebuilding the superblock and in any case we shouldn't assume
that fsck.gfs2 is running on the cluster the volume belongs to.

This patch removes the get_lockproto_table function and instead sets the
lock table name to a placeholder ("unknown") and sets lockproto to
"lock_dlm".  It warns the user at the end of the run that the locktable
will need to be set before mounting.

Signed-off-by: Andrew Price <anprice@redhat.com>