fsck.gfs2: Allocate enough space for the block map
Building with gcc 4.8's address sanitizer and running the test suite
throws up an issue where the block map allocated in fsck.gfs2 is not
large enough when the required size is an odd number. For example, when
the number of blocks is 3, the mapsize (number of chars to allocate
memory for) would be (3 >> 1) which is too small as each block requires
4 bits. This patch adds 1 to the mapsize, as suggested in an existing
comment, to ensure we have enough space in these situations.

Note that this has (probably) never made fsck.gfs2 misbehave as it only
ever accessed 4 bits past the end of the block map and by chance that
memory was never being modified by anything else in the rare cases in
which it was used.

Resolves: bz#947384

Signed-off-by: Andrew Price <anprice@redhat.com>