gfs2_lockcapture: Fixed header to use absolute path and modified some options
Fix the header to use absolute path. The "-o" has been modified
so that it takes a path that will be used to create the root directory of where
the data will be written and will be location of the tarball that is
created. Added undocumented(private) "-m" option that will capture the required
data and additional data that is not captured by default because sometimes it
can cause hangs to occur in certain situations. Removed the data capture of the
command "lsof" cause it might cause hangs when capturing lockdumps cause of
symlink lookup.

Signed-off-by: Shane Bradley <sbradley@redhat.com>