libgfs2: Add sd_heightsize bounds checking in read_sb
read_sb wasn't checking that x was less than the size of sd_heightsize
when looping over it. This patch adds a check for that. This resolves a
segfault in all tools which use read_sb, when the sb_bsize has been
zeroed in the superblock.

A test case has been added for this scenario in tests/fsck.at

Resolves: bz#1019226

Signed-off-by: Andrew Price <anprice@redhat.com>