gfs2_edit: Report referencing block address in the new journal code
When a block is referenced in the journal, the address of the "real"
referencing block is now printed (as well as the address of the block in
the journal).

Also, when a dinode references a block as a extended attribute, it
wasn't properly flagging it as a reference. Now it does.

Signed-off-by: Andrew Price <anprice@redhat.com>