libgfs2: Stick to the (rgrp) plan in lgfs2_rindex_entry_new
lgfs2_rindex_entry_new() was ignoring the plan any time len was not 0, even if
len matched the next rgrp length in the plan. This was causing problems in mkfs
when the journal-adjusted rgrp size was the same as the normal rgrp size.
Decrement the count of the appropriate rgrp size if the length matches.

Signed-off-by: Andrew Price <anprice@redhat.com>