libgfs2: Keep a pointer to the sbd in lgfs2_rgrps_t
As a set of resource groups is tied to a particular file system (block
size, device length...) it makes sense to keep a reference to the file
system in the lgfs2_rgrps_t type. This allows us to avoid duplication of
the bsize and device length fields, reduces parameter counts and
provides a convenient way to access file system-specific values in
resource group-related functions.

Signed-off-by: Andrew Price <anprice@redhat.com>