libgfs2: Remove debug field from gfs2_sbd
The utils were using the debug field to store their debug printing
options.  Hide the debug field in a libgfs2-internal header to allow
switching on libgfs2 debugging output separately and switch the apps to
use their own debug options. In cases where apps were using sdp->debug
but not actually setting it, remove the code which depends on it being
set.

Signed-off-by: Andrew Price <anprice@redhat.com>