gfs2_edit: Add a savemeta file metadata header
Previously 'gfs2_edit savemeta' just saved blocks into the file and any
characterisation of the contents had to be calculated by reading the
gfs2 structures out of it. Add a metadata header to savemeta output
files which initially holds the original fs size and a file creation
time, leaving some room to spare for future fields to be added.

Signed-off-by: Andrew Price <anprice@redhat.com>