gfs2_edit: Print block types with log descriptors
This patch makes gfs2_edit's journal print function print the metadata
block type along side the block number in its log descriptor output.
For example: In the past, when a log descriptor was printed, it looked
like this:
0x8200 (j+ 163): Log descriptor, type 300 (Metadata) len:504, data1: 503 (503)
                    0x70eb772          0x70df987          0x70dc7a1          0x70eb574
                    0x70dc5a3          0x70eb376          0x70dc3a5          0x70eb178
                    0x70dc1a7          0x70eaf7a          0x70dbfa9          0x70ead7c
With this patch, the output is more useful:
0x8200 (j+ 163): Log descriptor, type 300 (Metadata) len:504, data1: 503 (503)
                    0x70eb772 in       0x70df987 rb       0x70dc7a1 in       0x70eb574 in
                    0x70dc5a3 in       0x70eb376 in       0x70dc3a5 in       0x70eb178 in
                    0x70dc1a7 in       0x70eaf7a in       0x70dbfa9 in       0x70ead7c in
In this case, "in" indicates indirect blocks, and "rb" indicates a rgrp bitmap.
This is useful for getting a bigger picture of what that journal descriptor
contains and represents.