libgfs2: Write file metadata sequentially
Until now the journal creation functions built up the height of the
journal metadata using incremental allocation and the build_height
function which grows a file naively. Two things are about to change:

1. We will guarantee that a journal will occupy a single extent.
2. Journals will be written out sequentially and will only required one
   write of the resource group header, before the journal is written.

Since we know the size of the extent and can predict the layout of the
journal's metadata tree, it is possible to generate and write the
metadata blocks in sequence. This patch adds an lgfs2_write_filemeta()
function which does just that.

Signed-off-by: Andrew Price <anprice@redhat.com>