libgfs2: Create jindex directory separately from journals
Previously journals were created indirectly via build_jindex and the
jindex inode (and therefore the master inode) was created before the
journals. Now that we're allocating the journals in whole resource
groups at the start of the mkfs process we need a way to create the
jindex after the journals are created. This adds lgfs2_build_jindex
which takes a list of inums relating to journals and builds the jindex
from them.

The old build_jindex function is left in place until the other tools
have been switched to use the new function.

Signed-off-by: Andrew Price <anprice@redhat.com>