gfs2-utils tests: Build unit tests with consistent cpp flags
The check_rgrp unit test was failing on ARM and s390 due to struct
gfs2_sbd having a different alignment from libgfs2. This was due to the
unit tests being built with different options to libgfs2. Fix up the
unit test CPPFLAGS to make sure the alignment is consistent.

Also bump the test timeout for check_rgrp to make sure it completes on
slower build machines.

Signed-off-by: Andrew Price <anprice@redhat.com>