gfs2_edit: Use metadata description to print and assign fields
Replace the large number of separate field assignment and printing
functions in gfs2_edit with the new functions in libgfs2. This reduces
the amount of code involved dramatically and fixes three coverity
complaints related to the use of the 'checkassigns' macro.

This relates to the form: gfs2_edit -p 1234 field foo [bar] /dev/baz

The 'device' variable has been changed from char[NAME_MAX] to a pointer
to an argv member, so that now we can check whether the optional [bar]
argument is missing properly, i.e. no longer assume the device path will
begin with '/'. It also means the device name is no longer limited to
255 chars.

Signed-off-by: Andrew Price <anprice@redhat.com>