libgfs2: metafs_path improvements
Switch the metafs_path field in struct gfs2_sbd to be a pointer instead
of a char[PATH_MAX] and use strdup to set it in mount_gfs2_meta. Also
rework the metafs mounting and cleanup code to back out more robustly on
errors and set/reset signal handlers through a single function.

Note the SIGSEGV handler has been removed: when a SIGSEGV handler
returns, execution restarts at the bad instruction which triggered it,
so handling it could only ever cause an infinite loop.

Signed-off-by: Andrew Price <anprice@redhat.com>