gfs2_jadd: Don't use fixed size buffers for paths
This fixes some more PATH_MAX usages by dynamically allocating the space
to store metafs path strings. Previously the strings were rebuilt each
time they were required using sprintf and fixed size buffers. Remove the
sprintfs by storing the strings in the opts structure and pass around a
pointer to it instead of the gfs2_sbd where possible. This reduces the
number of uses of sbd.orig_journals which should be removed eventually
as it's application-specific.

Signed-off-by: Andrew Price <anprice@redhat.com>