fsck.gfs2: Reprocess nodes if anything changed - addendum 1 of 2
Before this patch, fsck.gfs2 called "reprocess_inode" in several
places, especially when the number of blocks had changed from the
original. That works in almost all cases. However, there's a corner
case where a block may be deleted, due to errors (for example, a
bad directory leaf), and another block takes its place. In that
case the count of the number of blocks is still the same, but the
inode should be reprocessed anyway, because the deleted blocks
and replacement blocks may be at different locations or maybe of
different types. A hash table block may be "data" when it's freed,
but if the block is reused, it may be repurposed as a "leaf" block.
That may confuse subsequent processing. Another reason to call
reprocess_inode is when the goal block changes to a value outside
the resource group, due to block allocations for repairs.

This patch is part 1 of 2 addendum fixes to make fsck fix i_goal
values correctly. The original three patches to fix this problem were
checked in against a wrong bz 1149516 (which is the rhel6 version
of the bug). The complete fix contains those three patches (b09dae8,
d79246d, c5b09c4) and this two-patch addendum for a total of 5
patches.

Resolves: rhbz#1178604
Signed-off-by: Bob Peterson <rpeterso@redhat.com>