fsck.gfs2: addendum to fix broken i_goal values in inodes - addendum 2 of 2
This patch moves some code around and fixes some corner cases that
the previous patches did not address.
This patch also fixes some trailing whitespace and removes a test
that is no longer valid from test/fsck.at

This patch is part 2 of 2 addendum fixes to make fsck fix i_goal
values correctly. The original three patches to fix this problem were
checked in against a wrong bz 1149516 (which is the rhel6 version
of the bug). The complete fix contains those three patches (b09dae8,
d79246d, c5b09c4) and this two-patch addendum for a total of 5
patches.

Resolves: rhbz#1178604
Signed-off-by: Abhi Das <adas@redhat.com>