gfs2_edit savemeta: Don't read rgrps twice
By the time savemeta() is called the rindex has already been read into
memory, and before savemeta() processes each resource group it calls
gfs2_rgrp_read() to read in the rgrp header. It frees the rgrp once it's
done with it.

Strangely, before processing the resource groups, savemeta() reads the
superblock a second time and calls ri_update() which reads the rindex
again, as well as reading in every rgrp header to be kept in memory for
the duration. This caused high memory usage and a noticeable performance
reduction when saving metadata of large file systems.

To solve these problems, this patch removes the code which re-reads the
superblock, rindex and the rgrps. The code which reads the rindex has
been reorganised for clarity and the sbd.fssize field is now set
properly at that point.

With this patch, using a large fs, I'm seeing improvements similar to:

Before: 43:36.12elapsed 10822268maxresident k
After:  28:44.67elapsed   226980maxresident k

Signed-off-by: Andrew Price <anprice@redhat.com>