gfs2_grow: Put back the definition of FALLOC_FL_KEEP_SIZE
An #include <linux/falloc.h> was removed in commit 14193bb. Although
fallocate(2) support was added to glibc 2.10, the FALLOC_FL_*
definitions were not added until 2.18. This means that gfs2-utils failed
to build on RHEL7 which has glibc 2.17. Add back the FALLOC_FL_KEEP_SIZE
definition in main_grow.c to fix that.

Signed-off-by: Andrew Price <anprice@redhat.com>