mkfs.gfs2: Allow longer cluster names
Increase the enforced limit for cluster name to 32 bytes and file
system name to 30 bytes for mkfs.gfs2 (was previously 16 + 16
bytes).

Also increased this limit in tunegfs2 when labelling gfs2 file
systems.

Updated the formation in the man pages along with adding a new test
case for mkfs.gfs2 to validate the increased cluster/file system
name support.

Resolves: rhbz#1202831
Signed-off-by: Paul Evans <pevans@redhat.com>