fsck.gfs2: Remove redundancy in add_duplicate_ref
Before this patch, when a duplicate reference was added to the list,
function add_duplicate_ref would print out an info message that said:
This brings the total to: X duplicate references.
I found this message to be redundant and confusing. It's much more
clear to simply say:
This brings the total to: X references